pyproject.toml
[project]
name = "scholar_digest_ai" # TODO: Update Package Name
version = "0.0.1"
authors = [
  { name = "harry pink", email = "hpink97@gmail.com" }
]
description = "" # TODO: Update Description
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "pdfplumber>=0.11.0",
  "langchain>=0.3.13",
  "sentence-transformers>=3.3.0",
  "chromadb>=0.5.23",
  "openai>=1.58.1",
  "python-dotenv>=1.0.0",
  "streamlit>=1.41.0",
  "emoji>=2.14.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["services*"]

[tool.black]
line-length = 110
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.venv
    | build
    | dist
    | migrations
  )/
)
'''

